Set ability to u16, implemented Unseen Fists
We can now have more than 255 abilities - first ability after that to be implemented is Unseen Fist by modifying "IsBattlerProtected" in battle_script_commands.c

+++ src/battle_message.c
@@ -50,7 +50,7 @@ static void ChooseTypeOfMoveUsedString(u8 *dst);
 static void ExpandBattleTextBuffPlaceholders(const u8 *src, u8 *dst);
 
 // EWRAM vars
-static EWRAM_DATA u8 sBattlerAbilities[MAX_BATTLERS_COUNT] = {0};
+static EWRAM_DATA u16 sBattlerAbilities[MAX_BATTLERS_COUNT] = {0};
 EWRAM_DATA struct BattleMsgData *gBattleMsgDataPtr = NULL;
 
 // const rom data
